backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
Pillow==10.3.0
numpy==1.26.4
torch==2.3.1
torchvision==0.18.1
opencv-python==4.9.0.80
timm==1.0.15